backend/requirements.txt
boto3==1.26.93
botocore==1.29.93
bs4==0.0.1
cmake==3.26.1
dj-database-url==1.2.0
Django==3.2.12
django-cors-headers==4.0.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
gunicorn==20.1.0
jsonfield==3.1.0
Pillow==9.1.0
psycopg2==2.9.5
PyJWT==2.7.0
pymongo==3.12.3
pyparsing==3.0.8
pywin32==303
razorpay==1.4.1
requests==2.27.1
sqlparse==0.2.4
uvicorn==0.20.0
whitenoise==6.4.0
zipp==3.8.0